moved historicalSearchMenu into SearchResults and modified its index store
